add predicted schedules
