+++ apps/site/assets/css/_schedule-page.scss
@@ -263,8 +263,8 @@
 }
 
 .schedule-table__row-route {
-  display: inline-flex;
   align-items: center;
+  display: inline-flex;
 }
 
 .schedule-table__row-route-pill {
