+++ apps/site/lib/site_web/controllers/mode_controller.ex
@@ -2,10 +2,11 @@ defmodule SiteWeb.ModeController do
   use SiteWeb, :controller
 
   alias Content.Teaser
+  alias PredictedSchedule.Display
+  alias PredictedSchedule.Schedules
   alias Site.TransitNearMe
   alias SiteWeb.{Mode, StopController}
   alias Routes.{Route}
-  alias PredictedSchedule.Display
   plug(SiteWeb.Plugs.RecentlyVisited)
   plug(SiteWeb.Plug.Mticket)
   plug(:require_google_maps)
@@ -45,7 +46,8 @@ defmodule SiteWeb.ModeController do
         "direction_id" => direction_id
       }) do
     schedules =
-      StopController.get_schedules(route_id, origin_stop, direction_id: direction_id)
+      route_id
+      |> Schedules.get_schedules(origin_stop, direction_id: direction_id)
       |> Enum.map(&TransitNearMe.build_time_map(&1, now: Util.now()))
       |> Enum.map(&route_with_prediction(&1, route_id))
       |> Enum.take(2)
